services/api-gateway/requirements.txt
Flask==3.1.2
flask-cors==5.0.1
kafka-python-ng==2.2.3
prometheus-client==0.21.1
gunicorn==25.3.0